+++ probe/process/reporter_test.go
@@ -10,10 +10,10 @@ import (
 )
 
 type mockWalker struct {
-	processes []*process.Process
+	processes []process.Process
 }
 
-func (m *mockWalker) Walk(f func(*process.Process)) error {
+func (m *mockWalker) Walk(f func(process.Process)) error {
 	for _, p := range m.processes {
 		f(p)
 	}
@@ -22,7 +22,7 @@ func (m *mockWalker) Walk(f func(*process.Process)) error {
 
 func TestReporter(t *testing.T) {
 	walker := &mockWalker{
-		processes: []*process.Process{
+		processes: []process.Process{
 			{PID: 1, PPID: 0, Comm: "init"},
 			{PID: 2, PPID: 1, Comm: "bash"},
 			{PID: 3, PPID: 1, Comm: "apache", Threads: 2},
